initial commit
This is a rough first commit for a modular synthesiser/music environment

The basic idea is that the programmer can define IoModules (structures
containing inputs, outputs, an oder and a processing function) in code,
then instantiateand connect these modules together during runtime, to
route audio signals through them.

A lot still needs to be planned/decided, but at the moment, the modules
are organised in a graph (hashmap), and modules are given a processing
order, wherein modules with an equal order should in theory be able to
be processed in parallel, as they will only rely on modules with a lower
order assigned.

Arc<RwLock<f64>> variables are used for ports, so that an output and input can
simply be "connected" via a shared variable. This is currently preferred
over mpsc::channel<f64>, as it allows a single output port to connect to multiple
input ports, without some need to duplicate and track data in between.

Currently modules are separate structs which share common funcitonality
via the IoMoudleTrait (name-change to follow) trait. These structs can
be defined via a declaritive macro. In the near future, I may just opt
for using a single IoModule struct, as the modules follow the same
standard structure, and only differ in the actual fields.