make ports directly accessible instead of via hashmap
This is done as accessing each port in the hashmap while processing is
too time consuming of a process.
This is fine for the trait-based approach, as a module just needs to
concern itself with its own unique set of ports and how to evaluate
them.
Also remove the IoModule struct entirely from the porject.