Begin work on an TUI interface
Add tui-rs and crossterm as dependencies and build a simple interface
for the synth which allows entering commands, printing their output and
displaying information on the rack's modules.

An new App struct has been created, which houses the new interface
controls and UI, in addition to the previous audio-thread setup and run
functions.