Add Control type
This type takes no input and is a relatively static value which requires
no processing. For this reason, it does not implement IoModule and is
stored and handled separately from the rack's modules and module chain.
Actual creation and handling could potentially be cleaner, but for now
it does what it needs to without slowing anything else up.