Change control to a trait and alter rack accordingly
There are a wide classification of things that could fit as a
Control, such as keyboards, sequencers, buttons, etc. For this reason,
Control is now a trait that is loosely defined as an unordered, no-input
module. The Rack now contains a separate hashmap of Control types, in
addition to a "focussed_module", which enables selecting a single
Control from the Rack's controls that will receive commands.

The previous "Control" struct is now "ControlKnob", which is more
specific to its purpose anyway.

A "Button" type which implements Control has also been added. It simply
toggles between 0 and 1 (deactivated and activated, respectively);