Set control_knob output to value if previously none
Upon control of a ControlKnob, if the value is None, set it to a value.
This preliminarily fixes a bug wherein a ControlKnob can't be adjusted
via a control, until it has been manually set with a value. A better fix
will be found when the structure is expanded upon.