adsr: fix logic errors in behaviour and cleanup committed
place a,d,s & r values in variablies, to avoid the long lock-read
statements.

make trigger_active a bool so that lock, read and comparisons to zero
musn't be present throughout the code (another long statement);

correct some minor logical issues in the output of behaviour of the
adsr. note: judging by the output, there are still some things to work
out, but it kind of works as intended..