fix set_ctrl_value function and call
The values being passed to this function were incorrectly placed, and
then the result of the operation ignored by the controller at the end.
Error handling is still needed throughout the call stack, but now the
"set" command works as intended and actually sets the value, rather than
silently failing. The TODO comment was also removed as this function is
still quite convenient for testing