Set default adsr time values to clock.time_delta
In order to give them a value that's basically zero, without actually
being zero.

Note, that although the zero division stated in the comment currently
isn't a problem (due to updating the active_time before transitions), I
wanted to make note of it and set a good default in case of change in
the active_time update behaviour.