Remove ref/lock on clock's time field
And pass clock reference to modules (oscillator), instead of time.

The clock itself is already reference counted for synchronisation. It's
easier to just pass a reference of the entire clock to the modules that
need it, instead of further having to lock one of it's values.

Also, chnange SAMPLE_RATE to be f32 as it's only used with f32 values