Use f64 instead of f32 in yat-rack
It only needs to be f32 for the output sample, so can be converted at
this point (basically a truncated amplitude with more than enough
resolution).

In the actual audio processing phase, particularly for clock
incrementation, the increased resolution could be better for quality, at
no extra cost (in theory).