make clock's time count to 100,000
This address the problem of audible clicks when the clock's time field
wraps back to zero.

It doesn't technically fix the problem, but having it occur just once
every ~27 hours seems reasonable for now.

Note the value of 100,000 is relatively arbitrary, outside of being
small enough to retain resolution while incrementing the time value.