adsr: keep track of amplitude at release time
This is used to smoothly transition from whatever state the ADSR is in
to the release state.
Prior to this, the release amplitude would start from the sustain
amplitude level, meaning that a key release in any other state would
cause a sudden increase/decrease in amplitude, prior to the fade out.

The 'time' argument that was passed to the constuctor has been renamed
to 'clock' in accordance with the the internal type and field name.