adsr: don't process an input signal
An ADSR shouldn't process an input signal, but rather just output a
signal from 0 to 1, based on time and a gate input.

This patch removes the audio input from the envelope and renames
audio_out to signal_out and trigger_in to gate_in for clarity.

Functionally it's similar, however, this is both simpler and more
correct.